backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
tensorflow==2.19.0
Pillow==11.2.1
numpy==2.1.3
